Refactor case search to replace `stringId` with `id`.
Replaced the deprecated `stringId` field with the new `id` field throughout the codebase, maintaining backward compatibility by merging both fields during queries. Updated `ImmediateField` and its references to simplify type handling, and moved it to a more appropriate package. These changes improve clarity, support future deprecation, and ensure smooth transitions.
